+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New checks:
 * ``empty-background``
 * ``outline-could-be-a-scenario``
 * ``consider-using-background``
+* ``examples-outside-scenario-outline``
 
 Other changes:
 * ``unparseable-file`` is now issued for all nodes that offend the ``gherkin`` parser together with more error information provided by it